Time: 0 ms (100%), Space: 77.1 MB (75.41%) - LeetHub

+++ 1701-average-waiting-time/1701-average-waiting-time.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     double averageWaitingTime(vector<vector<int>>& customers) {
-        float waiting = 0;
+        double waiting = 0;
         int last_finish = customers[0][0];
         for (int i = 0; i < customers.size(); i++){
             int extra_time = 0;
